skills/ai-crypto/xspoonai-official-fastkol-web3-content-pipeline/scripts/spoon_ai_stub.py
import asyncio
from typing import Dict, Any, List, Optional
from pydantic import BaseModel

class ToolManager:
    def __init__(self, tools: List[Any]):
        self.tools = {t.name: t for t in tools}

    def get_tool(self, name: str):
        return self.tools.get(name)

class MCPTool:
    def __init__(self, name: str, description: str, mcp_config: Dict[str, Any]):
        self.name = name
        self.description = description
        self.mcp_config = mcp_config

    async def execute(self, **kwargs) -> Any:
        raise NotImplementedError(f"Tool {self.name} execution not implemented in base class.")

class ChatBot:
    def __init__(self, model_name: str, temperature: float = 0.7, max_tokens: int = 4096):
        self.model_name = model_name
        self.temperature = temperature
        self.max_tokens = max_tokens

    async def generate_response(self, system_prompt: str, user_content: str) -> str:
        return "This is a LLM response."

class SpoonReactMCP:
    def __init__(self, tools: List[Any] = None):
        self.tools = tools or []
        self.tool_manager = ToolManager(self.tools)

    async def run(self, input_text: str):
        pass
