skills/ai-crypto/xspoonai-official-fastkol-web3-content-pipeline/scripts/requirements.txt
httpx
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
py-clob-client
python-dotenv
